requirements.txt
gradio==6.0.0
Pillow==12.0.0
ultralytics==8.3.232
